+++ network.cpp
@@ -6,7 +6,7 @@
 #include <iostream>
 #include "../../../../../../Program Files (x86)/Microsoft Visual Studio/2019/Enterprise/VC/Tools/MSVC/14.21.27702/include/ctime"
 
-Network::Network(Layer** layers, unsigned long layerCount, double learningRate, unsigned short batchSize, float momentum,
+Network::Network(Layer** layers, size_t layerCount, double learningRate, unsigned short batchSize, float momentum,
 	bool init) {
 	srand(time(NULL));
 
@@ -16,7 +16,7 @@ Network::Network(Layer** layers, unsigned long layerCount, double learningRate, 
 	this->momentum = momentum;
 	this->layerCount = layerCount;
 
-	for (unsigned long i = 1; i < layerCount; i++) {
+	for (size_t i = 1; i < layerCount; i++) {
 		layers[i]->weights = (new Matrix(layers[i]->nodeCount, layers[i - 1]->nodeCount))->random();
 		layers[i]->previousWeight = (new Matrix(layers[i]->nodeCount, layers[i - 1]->nodeCount, true));
 		layers[i]->deltaWeightSum = Matrix::copy(layers[i]->previousWeight);
@@ -32,18 +32,19 @@ Network::~Network()
 	// delete[] layers;
 }
 
-double* Network::predict(double* inputs, unsigned long length)
+double* Network::predict(double* inputs, size_t length)
 {
-	if (length != layers[0]->nodeCount) { return nullptr; }
+	if (length != layers[0]->nodeCount) {
+		return nullptr;
+	}
 
 	/* Set the input layer neurons to the passed inputs */
-	for (size_t i = 0; i < length; ++i)
-	{
+	for (size_t i = 0; i < length; ++i) {
 		layers[0]->neurons[i] = inputs[i];
 	}
 
 	/* Feed forward algorithm */
-	for (unsigned long i = 1; i < layerCount; i++) {
+	for (size_t i = 1; i < layerCount; i++) {
 		/*
 		 * 1. Multiply previous Neurons with current index I weights to get new Neuron Matrix for index I
 		 * 2. add() bias to calclulated neurons
@@ -62,12 +63,12 @@ double* Network::predict(double* inputs, unsigned long length)
 }
 
 void Network::printLastResult() {
-	for (unsigned long i = 0; i < layers[layerCount-1]->nodeCount; i++) {
+	for (size_t i = 0; i < layers[layerCount-1]->nodeCount; i++) {
 		std::cout << layers[layerCount - 1]->neurons[i] << std::endl;
 	}
 }
 
-void Network::train(double* input, unsigned long inputLength, double* targetOutput, unsigned long targetOutputLength) {
+void Network::train(double* input, size_t inputLength, double* targetOutput,size_t targetOutputLength) {
 	if (targetOutputLength != layers[layerCount - 1]->nodeCount) { 
 		std::cout << "wrong training data";
 		return; 
@@ -81,8 +82,7 @@ void Network::train(double* input, unsigned long inputLength, double* targetOutp
 	errorMatrices[layerCount - 1] = outputError;
 
 	/* Iterating through layers from behind */
-	for (unsigned long i = layerCount - 1; i > 0; i--)
-	{
+	for (size_t i = layerCount - 1; i > 0; i--) {
 		double* neuronsDerivative = ArrayHelper::mapTo(layers[i]->neurons, layers[i]->nodeCount, layers[i]->func, true);
 
 		double* errorMultNeurons = ArrayHelper::hadamardArray(neuronsDerivative, errorMatrices[i], layers[i]->nodeCount);
@@ -106,8 +106,7 @@ void Network::train(double* input, unsigned long inputLength, double* targetOutp
 		delete errorMatrix;
 	}
 
-	for (size_t i = 0; i < layerCount; ++i)
-	{
+	for (size_t i = 0; i < layerCount; ++i) {
 		delete[] errorMatrices[i];
  	}
 	delete[] errorMatrices;
